Move tag fetch utils to `cmd/internal` package
