+++ cmd/krew/cmd/internal/fetch_tag.go
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package updatecheck
+package internal
 
 import (
 	"encoding/json"
